Seems like I completed flags management. Also, fixed some functions in libft

+++ libft/ft_strnstr.c
@@ -14,7 +14,7 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int		beg;
+	size_t	beg;
 	size_t	i;
 	size_t	j;
